services/auth/requirements/install.txt
# =============================================================
# requirements/install.txt -- auth service (pip format)
# =============================================================

fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
psycopg2-binary==2.9.10
python-multipart==0.0.6
tzdata==2026.1
